[W5.3b] LO-Inheritance +Added Contact class +Modifed Phone, Address, Contact class to extends contact class
