+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,32 @@
+package seedu.addressbook.data.person;
+
+public class Contact {
+    public final String value;
+    private boolean isPrivate;
+
+    public Contact (String value, boolean isPrivate){
+        this.value = value.trim();
+        this.isPrivate = isPrivate;
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this // short circuit if same object
+                || (other instanceof Email // instanceof handles nulls
+                && this.value.equals(((Email) other).value)); // state check
+    }
+
+    @Override
+    public String toString() {
+        return value;
+    }
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+
+}
